added di container, swagger, logger, create\get methods

+++ DeviceMessagesConsumer/DataAccess/Entities/Device.cs
@@ -4,10 +4,9 @@ namespace DeviceMessagesConsumer.DataAccess.Entities
 {
     public class Device
     {
-        public long Id { get; set; }
+        public short Id { get; set; }
         
         public string Name { get; set; }
-        
         public bool IsActive { get; set; }
         
         public ICollection<Measurement> Measurements { get; set; }
